requirements.txt
Pillow==12.1.1
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.54.0
